pyproject.toml
[tool.poetry]
name = "asset-insight-graph"
version = "0.1.0"
description = "GraphRAG example with Neo4j"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "api"}, {include = "etl"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
neo4j = "^5.14"
fastapi = "^0.110"
uvicorn = {extras = ["standard"], version = "^0.27"}
python-dotenv = "^1.0"
pydantic-settings = "^2.2"
langchain = "^0.1.14"
openai = "^1.12"
pandas = "^2.1"
tqdm = "^4.66"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.23"
black = "^23.9"
isort = "^5.12"
ruff = "^0.1.14"
pre-commit = "^3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
